+++ lua/system/config.lua
@@ -1,3 +1,5 @@
+-- norns configuration
+
 -- add some stuff to package.path
 -- this consists of search patterns lua will use for require('foo')
 
@@ -7,6 +9,7 @@ local sys = norns..'/?.lua;'
 local lib = home..'/dust/lib/lua/?.lua;'
 
 package.path = sys..lib..package.path
+-- print('package.path: ' .. package.path)
 
 -- full path to directory containing user scripts
 script_dir = home..'/dust/scripts/'
@@ -14,6 +17,3 @@ data_dir = home..'/dust/data/'
 audio_dir = home..'/dust/audio/'
 home_dir = home
 
-require 'norns'
-
--- print('package.path: ' .. package.path)
