+++ lua/system/lib/script.lua
